refactor(reader): consolidate end page metrics
Merge platform-specific end page layout metrics into a single cross-platform helper.

Add PlatformFont beside PlatformImage so native end page typography follows the existing platform alias pattern.
